+++ pkg/ui/tui/main.go
@@ -2,60 +2,26 @@ package tui
 
 import (
 	"fmt"
-	"github.com/ZaparooProject/zaparoo-core/pkg/api/client"
 	"github.com/ZaparooProject/zaparoo-core/pkg/config"
 	"github.com/ZaparooProject/zaparoo-core/pkg/platforms"
 	"github.com/ZaparooProject/zaparoo-core/pkg/utils"
 	"github.com/gdamore/tcell/v2"
 	"github.com/rivo/tview"
-	"github.com/rs/zerolog/log"
-	"os/exec"
-	"path"
-	"strings"
 	"time"
 )
 
-func setupLogExport(
-	pl platforms.Platform,
-	app *tview.Application,
-	pages *tview.Pages,
-	logDestPath string,
-	logDestName string,
-) *tview.List {
-	logExport := tview.NewList()
-
-	logExport.
-		AddItem("Upload to termbin.com", "", 'a', func() {
-			pages.RemovePage("export")
-			outcome := uploadLog(pl, pages, app)
-			modal := genericModal(outcome, "Log upload", func(buttonIndex int, buttonLabel string) {
-				pages.RemovePage("upload")
-			}, true)
-			pages.AddPage("upload", modal, true, true)
-		})
-
-	if logDestPath != "" {
-		logExport.AddItem("Copy to "+logDestName, "", 'b', func() {
-			pages.RemovePage("export")
-			outcome := copyLogToSd(pl, logDestPath, logDestName)
-			modal := genericModal(outcome, "Log copy", func(buttonIndex int, buttonLabel string) {
-				pages.RemovePage("copy")
-			}, true)
-			pages.AddPage("copy", modal, true, true)
-		})
-	}
-
-	logExport.AddItem("Cancel", "", 'q', func() {
-		pages.RemovePage("export")
-	}).ShowSecondaryText(false)
-
-	logExport.
-		SetBorder(true).
-		SetBorderPadding(1, 1, 1, 1).
-		SetTitle("Log export")
-
-	return logExport
-}
+const (
+	PageMain              = "main"
+	PageSettingsMain      = "settings_main"
+	PageSettingsTags      = "settings_tags"
+	PageSettingsTagsRead  = "settings_tags_read"
+	PageSettingsTagsWrite = "settings_tags_write"
+	PageSettingsAudio     = "settings_audio"
+	PageSettingsReaders   = "settings_readers"
+	PageSettingsScanMode  = "settings_readers_scanMode"
+	PageSearchMedia       = "search_media"
+	PageExportLog         = "export_log"
+)
 
 func BuildMain(
 	cfg *config.Instance,
@@ -111,7 +77,7 @@ func BuildMain(
 	displayCol.AddItem(helpText, 3, 1, false)
 
 	pages := tview.NewPages().
-		AddPage("main", main, true, true)
+		AddPage(PageMain, main, true, true)
 
 	// create the main modal
 	main.SetTitle("Zaparoo Core v" + config.AppVersion + " (" + pl.ID() + ")").
@@ -121,10 +87,10 @@ func BuildMain(
 	main.AddItem(displayCol, 0, 1, false)
 
 	searchButton := tview.NewButton("Search media").SetSelectedFunc(func() {
-		app.Stop()
+		pages.SwitchToPage(PageSearchMedia)
 	})
 	searchButton.SetFocusFunc(func() {
-		helpText.SetText("Search for media and write to a tag.")
+		helpText.SetText("Search for media and write to an NFC tag.")
 	})
 
 	updateDBButton := tview.NewButton("Update media DB").SetSelectedFunc(func() {
@@ -135,23 +101,15 @@ func BuildMain(
 	})
 
 	settingsButton := tview.NewButton("Settings").SetSelectedFunc(func() {
-		enableZapScript := client.DisableZapScript(cfg)
-		_, err := ConfigUiBuilder(cfg, app, pages, func() {
-			enableZapScript()
-			pages.SwitchToPage("main")
-		})
-		if err != nil {
-			log.Error().Err(err).Msg("error running config app")
-			app.Stop()
-		}
+		pages.SwitchToPage(PageSettingsMain)
 	})
 	settingsButton.SetFocusFunc(func() {
 		helpText.SetText("Manage settings for Core service.")
 	})
 
 	exportButton := tview.NewButton("Export log").SetSelectedFunc(func() {
-		widget := modalBuilder(setupLogExport(pl, app, pages, logDestPath, logDestName), 42, 8)
-		pages.AddPage("export", widget, true, true)
+		widget := modalBuilder(BuildExportLog(pl, app, pages, logDestPath, logDestName), 42, 8)
+		pages.AddPage(PageExportLog, widget, true, true)
 	})
 	exportButton.SetFocusFunc(func() {
 		helpText.SetText("Export Core log file for support.")
@@ -237,38 +195,16 @@ func BuildMain(
 
 	main.AddItem(buttonRowTop, 20, 1, true)
 
+	BuildSettingsMainMenu(cfg, pages, app)
+	BuildTagsMenu(cfg, pages, app)
+	BuildTagsReadMenu(cfg, pages, app)
+	BuildSearchMedia(cfg, pages, app)
+	BuildTagsWriteMenu(cfg, pages, app)
+	BuildAudioMenu(cfg, pages, app)
+	BuildReadersMenu(cfg, pages, app)
+	BuildScanModeMenu(cfg, pages, app)
+	pages.SwitchToPage(PageMain)
+
 	centeredPages := centerWidget(70, 20, pages)
 	return app.SetRoot(centeredPages, true).EnableMouse(true), nil
-}
-
-func copyLogToSd(pl platforms.Platform, logDestPath string, logDestName string) string {
-	logPath := path.Join(pl.Settings().TempDir, config.LogFile)
-	newPath := logDestPath
-	err := utils.CopyFile(logPath, newPath)
-	outcome := ""
-	if err != nil {
-		outcome = fmt.Sprintf("Unable to copy log file to %s.", logDestName)
-		log.Error().Err(err).Msgf("error copying log file")
-	} else {
-		outcome = fmt.Sprintf("Copied %s to %s.", config.LogFile, logDestName)
-	}
-	return outcome
-}
-
-func uploadLog(pl platforms.Platform, pages *tview.Pages, app *tview.Application) string {
-	logPath := path.Join(pl.Settings().TempDir, config.LogFile)
-	modal := genericModal("Uploading log file...", "Log upload", func(buttonIndex int, buttonLabel string) {}, false)
-	pages.RemovePage("export")
-	// FIXME: this is not updating, too busy
-	pages.AddPage("temp_upload", modal, true, true)
-	app.ForceDraw()
-	uploadCmd := "cat '" + logPath + "' | nc termbin.com 9999"
-	out, err := exec.Command("bash", "-c", uploadCmd).Output()
-	pages.RemovePage("temp_upload")
-	if err != nil {
-		log.Error().Err(err).Msgf("error uploading log file to termbin")
-		return "Unable to upload log file."
-	} else {
-		return "Log file URL:\n" + strings.TrimSpace(string(out))
-	}
 }
